requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
alembic
python-jose
passlib[bcrypt]
celery[redis]
redis
